pyproject.toml
[project]
name = "task-management-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-debug-toolbar>=6.1.0",
    "django-environ>=0.12.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "drf-yasg>=1.21.11",
    "environ>=1.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pre-commit>=4.4.0",
    "psutil>=7.1.3",
    "psycopg[binary,pool]>=3.2.12",
    "redis>=7.1.0",
    "uvicorn[standard]>=0.22.0",
    "gunicorn>=20.1.0",
    "whitenoise>=6.5.0",
]
